+++ client/src/components/Chart/ChartPage.jsx
@@ -8,43 +8,44 @@ import { useLocation } from "react-router-dom";
 import ChartForm from "./chartForm/ChartForm";
 import LineChart from "./LineChart/LineChart";
 import { GET } from "../../constants/actionTypes";
-import newQuery from "./chartForm/chartQuery/chartQueryClass";
+import newQuery from "./chartForm/chartQuery/chartQueryFunctions";
 
 // eslint-disable-next-line no-unused-vars
 function ChartPage({ currentCamera, setCurrentCamera }) {
   const location = useLocation();
-  // const history = useHistory();
-
   const params = location.search ? location.search : null;
 
-  /* *************** Component State *************** */
-  // eslint-disable-next-line no-unused-vars
+  /* ____________________ Component State ____________________ */
   const [loading, setLoading] = useState(false);
   // eslint-disable-next-line no-unused-vars
   const [vehicleData, setVehicleData] = useState([]);
-  // eslint-disable-next-line no-unused-vars
   const [chartFilters, setChartFilters] = useState({});
-
-  // eslint-disable-next-line no-unused-vars
   const [urlQuery, setUrlQuery] = useState("");
 
-  // create an effect for updating the urlQuery object when the ChartFilters object changes
-  // useEffect()
-
-  // effects for when the API is called
   useEffect(() => {
+    /**
+     * Sets the UrlQuery when the chartFilters object changes, which only changes when the "submit" button on the charts page is clicked
+     */
+    if (
+      chartFilters.combinedDates != null &&
+      chartFilters.combinedDates !== undefined &&
+      chartFilters.querySelectedVehicles != null &&
+      chartFilters.querySelectedVehicles !== undefined
+    ) {
+      setUrlQuery(newQuery(chartFilters));
+    }
+  }, [chartFilters]);
+
+  useEffect(() => {
+    /**
+     * Makes the API call to the Backend and returns the data for the chart when urlQuery is changed
+     */
     // https://hmos.dev/en/how-to-cancel-at-axios
     const cancelToken = axios.CancelToken;
     const source = cancelToken.source();
 
     setLoading(true);
     const fetchData = async () => {
-      // setUrlQuery(chartQueryCreator(chartFilters));
-      console.log("FROM CHART PAGE", chartFilters);
-      // const chartQuery2 = new ChartQuery(chartFilters);
-      // console.log("NEW QUERY", chartQuery2.newQuery());
-      const newQString = newQuery(chartFilters);
-      setUrlQuery(newQString);
       console.log("new Query", urlQuery);
       try {
         let query;
@@ -57,10 +58,8 @@ function ChartPage({ currentCamera, setCurrentCamera }) {
           method: GET,
           url: `/vehicles?${query}`,
           cancelToken: source.token,
-          // cancelToken: new axios.CancelToken((c) => (cancel = c)),
         });
         setVehicleData(data.data);
-        console.log(urlQuery);
         setLoading(false);
       } catch (error) {
         if (axios.isCancel(error)) {
@@ -70,10 +69,7 @@ function ChartPage({ currentCamera, setCurrentCamera }) {
       }
     };
     fetchData();
-
-    // return () => cancel();
-    // When chartFilters changes, call the API again
-  }, [chartFilters, params]);
+  }, [urlQuery]);
 
   return (
     <Grid
@@ -85,16 +81,6 @@ function ChartPage({ currentCamera, setCurrentCamera }) {
     >
       <Grid.Col md={2.5} lg={2.5}>
         <ChartForm setChartFilters={setChartFilters} />
-        {/* <ChartForm
-              chartFilters={chartFilters}
-              setChartFilters={setChartFilters}
-              loading={loading}
-              urlQuery={urlQuery}
-              setUrlQuery={setUrlQuery}
-              navigate={navigate}
-              currentCamera={currentCamera}
-              setCurrentCamera={setCurrentCamera}
-            /> */}
       </Grid.Col>
       <Grid.Col md={9.5} lg={9.5}>
         <LineChart
@@ -109,4 +95,3 @@ function ChartPage({ currentCamera, setCurrentCamera }) {
 }
 
 export default ChartPage;
-// https://yarnpkg.com/package/react-charts#readme
