+++ client/src/components/Chart/chartForm/chartFormInputElements/rangeSlider.jsx
