--- client/src/components/Chart/chartForm/chartVariables.js
@@ -1,7 +1,0 @@
-const timeOptions = ["Past Hour", "Past Day", "Past Week", "Past Month"];
-const vehicleOptions = ["All", "Commuter", "Truck", "Bus", "Motorcycle"];
-const laneOptions = ["All", "1", "2", "3", "4"];
-const speedOptions = ["All", "0-50", "50-60", "60-70", "70+"];
-const tempOptions = ["All", "0-32", "32-60", "60-80", "70-90", "80-100"];
-
-export { timeOptions, vehicleOptions, laneOptions, speedOptions, tempOptions };
