--- client/src/components/Chart/styles.js
@@ -1,31 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-export default makeStyles(() => ({
-  // root: {
-  //   "& .MuiTextField-root": {
-  //     margin: theme.spacing(1),
-  //   },
-  // },
-  // paper: {
-  //   padding: theme.spacing(2),
-  // },
-  // form: {
-  //   display: "flex",
-  //   flexWrap: "wrap",
-  //   justifyContent: "center",
-  // },
-  // fileInput: {
-  //   width: "97%",
-  //   margin: "10px 0",
-  // },
-  // buttonSubmit: {
-  //   marginBottom: 10,
-  // },
-  lineChart: {
-    background: "white",
-    // height: "100%",
-    height: "max-content",
-    width: "100%",
-    // item: { display: "flex" },
-  },
-}));
